Refactor groupId
GroupId is also the deployment path. These topics are part of drb,
and cortex, so fr.gael.drb.cortex.sentinel-2 groupId is preferred
to fr.gael.cortex. This topic splits sentinel-2 into 4 topics that
must be gathered into a "sentinel-2" container.

In addition names of the packages has been set.